battery_info: remove 4K stack buffer
We have an Arena for this purpose and we can determine the needed
size at compile time so we don't need to do this.

+++ config.def.h
@@ -14,7 +14,7 @@
 // static struct vol_arg va = { "default", "Speaker" };
 
 /* check blocks/xxx/battery_info.c for info */
-static struct bat_arg ba  = {.bat = "BAT0", .interval = 30};
+static struct bat_arg ba  = {.bat = s8("BAT0"), .interval = 30};
 
 /* check blocks/date.c for info */
 static struct date_arg da = {.fmt = "%R", .interval = 30};
